+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -2,11 +2,7 @@ package roomescape.controller;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.ReservationTime;
 import roomescape.service.ReservationTimeService;
@@ -22,16 +18,5 @@ public class ReservationTimeController {
     @GetMapping("/times")
     public ResponseEntity<List<ReservationTime>> read() {
         return ResponseEntity.ok(reservationTimeService.findAll());
-    }
-
-    @PostMapping("/times")
-    public ResponseEntity<ReservationTime> create(@RequestBody ReservationTime reservationTime) {
-        return ResponseEntity.ok(reservationTimeService.save(reservationTime));
-    }
-
-    @DeleteMapping("/times/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-        reservationTimeService.delete(id);
-        return ResponseEntity.ok().build();
     }
 }
